refactor(formatter,utils,compiler): move buildTokenRegex to utils & streamline token handling
- Moved `buildTokenRegex` from `compiler.js` to `utils.js` and updated all imports.
- Hoisted `tokenRE` and `defaultTokenKeys` in `formatter.js` to eliminate duplicate calls.
- Cloned `overrideTokens` into `userOverride` to avoid mutating caller objects.
- Consolidated `validTokens` and `produceable` sets to use `defaultTokenKeys`.
