+++ src/formatter.js
@@ -1,6 +1,7 @@
 import { DEFAULT_HANDLERS } from './handlers.js';
 import { parseInput } from './parser.js';
-import { compileFormat, buildTokenRegex } from './compiler.js';
+import { compileFormat } from './compiler.js';
+import { buildTokenRegex } from './utils.js';
 
 /**
  * formatDate()
@@ -20,10 +21,8 @@ export function formatDate(
 ) {
   if (typeof inputDate !== 'string') return null;
 
-  // 1) Merge handlers (built‑ins + custom)
   const handlers = { ...DEFAULT_HANDLERS, ...customTokens };
 
-  // 2) Parse (strict or best‑effort)
   let dateParts;
   try {
     dateParts = parseInput(inputDate, inputFormat, handlers, yearConverter);
@@ -35,60 +34,57 @@ export function formatDate(
     throw err;
   }
 
-  // 3) Derive which tokens were in the inputFormat
   const tokenRE = buildTokenRegex(handlers);
+  const defaultTokenKeys = Object.keys(defaultTokens);
   const parsedTokens = [...inputFormat.matchAll(tokenRE)].map((m) => m[0]);
+
   // allow yyyy output if input had yy + converter
   if (parsedTokens.includes('yy') && typeof yearConverter === 'function') {
     parsedTokens.push('yyyy');
   }
   dateParts._tokens = parsedTokens;
 
-  // 3a) Promote missing defaultTokens into overrideTokens so they fire first
-  for (const tok of Object.keys(defaultTokens)) {
+  const userOverride = { ...overrideTokens };
+  for (const tok of defaultTokenKeys) {
     if (!parsedTokens.includes(tok)) {
-      overrideTokens[tok] = defaultTokens[tok];
+      userOverride[tok] = defaultTokens[tok];
     }
   }
 
-  // 5) Build validity sets
   const validTokens = new Set([
     ...Object.keys(handlers),
-    ...Object.keys(defaultTokens),
+    ...defaultTokenKeys,
     ...Object.keys(overrideTokens),
   ]);
+
   const produceable = new Set([
     ...Object.keys(overrideTokens),
     ...parsedTokens,
-    ...Object.keys(defaultTokens),
+    ...defaultTokenKeys,
     ...Object.keys(customTokens),
   ]);
 
-  // 6) Pre‑validate & inject literal overrides in silent mode
   const needed = [...outputFormat.matchAll(tokenRE)].map((m) => m[0]);
   for (const tok of needed) {
     if (!validTokens.has(tok)) {
       if (errorPolicy === 'throw') {
         throw new Error(`Unknown token "${tok}" in output format`);
       }
-      // silent: literal‑fallback
-      overrideTokens[tok] = tok;
+      userOverride[tok] = tok;
       continue;
     }
     if (!produceable.has(tok)) {
       if (errorPolicy === 'throw') {
         throw new Error(`Cannot produce token "${tok}"—no data or default`);
       }
-      // silent: literal‑fallback
-      overrideTokens[tok] = tok;
+      userOverride[tok] = tok;
     }
   }
 
-  // 7) Compile & render
   const template = compileFormat(
     outputFormat,
     handlers,
-    overrideTokens,
+    userOverride,
     defaultTokens
   );
   return template
